# branch-1/swagger_docs.yml
swagger: '2.0'
info:
  title: Sistema de Ticketing con SQLAlchemy
  description: API REST para gestión de tickets con relación 1:N (Ticket → Incidentes)
  version: 2.0.0
tags:
  - name: "Tickets"
    description: "Operaciones relacionadas con tickets (1:N con incidentes)"
  - name: "Incidentes"
    description: "Operaciones relacionadas con incidentes (N:1 con ticket)"

paths:
  /tickets:
    get:
      tags:
        - "Tickets"
      summary: "Obtiene la lista de todos los tickets."
      parameters:
        - name: "incluir_incidentes"
          in: "query"
          type: "boolean"
          required: false
          description: "Incluir lista de incidentes asociados"
      responses:
        200:
          description: "Lista de tickets obtenida exitosamente"
    post:
      tags:
        - "Tickets"
      summary: "Crea un nuevo ticket con incidentes asociados (relación 1:N)."
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              cliente_id:
                type: "integer"
                example: 1
              servicio_id:
                type: "integer"
                example: 1
              equipo_id:
                type: "integer"
                example: 1
              empleado_id:
                type: "integer"
                example: 1
              incidentes:
                type: "array"
                description: "Lista de incidentes asociados al ticket"
                items:
                  type: "object"
                  properties:
                    descripcion:
                      type: "string"
                      example: "Pantalla no enciende"
                    categoria:
                      type: "string"
                      enum: ["Hardware", "Software", "Red", "Otro"]
                      example: "Hardware"
                    prioridad:
                      type: "string"
                      enum: ["Baja", "Media", "Alta", "Crítica"]
                      example: "Alta"
            required:
              - "cliente_id"
              - "servicio_id"
              - "equipo_id"
              - "empleado_id"
      responses:
        201:
          description: "Ticket creado exitosamente"
        400:
          description: "Parámetros inválidos"

  /tickets/{ticket_id}:
    get:
      tags:
        - "Tickets"
      summary: "Obtiene los detalles de un ticket con sus incidentes."
      parameters:
        - name: "ticket_id"
          in: "path"
          type: "integer"
          required: true
        - name: "incluir_incidentes"
          in: "query"
          type: "boolean"
          required: false
          default: true
      responses:
        200:
          description: "Ticket encontrado"
        404:
          description: "Ticket no encontrado"

  /tickets/{ticket_id}/incidentes:
    post:
      tags:
        - "Tickets"
      summary: "Agrega un nuevo incidente a un ticket existente."
      parameters:
        - name: "ticket_id"
          in: "path"
          type: "integer"
          required: true
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              descripcion:
                type: "string"
                example: "Teclado no responde"
              categoria:
                type: "string"
                enum: ["Hardware", "Software", "Red", "Otro"]
              prioridad:
                type: "string"
                enum: ["Baja", "Media", "Alta", "Crítica"]
            required:
              - "descripcion"
              - "categoria"
              - "prioridad"
      responses:
        201:
          description: "Incidente agregado exitosamente"
        400:
          description: "Parámetros inválidos"
        404:
          description: "Ticket no encontrado"

  /tickets/{ticket_id}/estado:
    put:
      tags:
        - "Tickets"
      summary: "Cambia el estado de un ticket."
      parameters:
        - name: "ticket_id"
          in: "path"
          type: "integer"
          required: true
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              estado:
                type: "string"
                enum: ["Abierto", "En Progreso", "Cerrado", "Reabierto"]
            required:
              - "estado"
      responses:
        200:
          description: "Estado actualizado exitosamente"
        400:
          description: "Estado inválido"

  /tickets/{ticket_id}/cerrar:
    put:
      tags:
        - "Tickets"
      summary: "Cierra un ticket registrando la fecha de cierre."
      parameters:
        - name: "ticket_id"
          in: "path"
          type: "integer"
          required: true
      responses:
        200:
          description: "Ticket cerrado exitosamente"
        404:
          description: "Ticket no encontrado"

  /tickets/{ticket_id}/reabrir:
    put:
      tags:
        - "Tickets"
      summary: "Reabre un ticket cerrado."
      parameters:
        - name: "ticket_id"
          in: "path"
          type: "integer"
          required: true
      responses:
        200:
          description: "Ticket reabierto exitosamente"
        400:
          description: "No se pudo reabrir"

  /tickets/filtrar/estado:
    get:
      tags:
        - "Tickets"
      summary: "Filtra tickets por estado."
      parameters:
        - name: "estado"
          in: "query"
          type: "string"
          required: true
          enum: ["Abierto", "En Progreso", "Cerrado", "Reabierto"]
      responses:
        200:
          description: "Tickets filtrados exitosamente"

  /incidentes:
    get:
      tags:
        - "Incidentes"
      summary: "Obtiene la lista de todos los incidentes."
      responses:
        200:
          description: "Lista de incidentes obtenida exitosamente"
    post:
      tags:
        - "Incidentes"
      summary: "Crea un nuevo incidente asociado a un ticket."
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              descripcion:
                type: "string"
                example: "Pantalla no enciende"
              categoria:
                type: "string"
                enum: ["Hardware", "Software", "Red", "Otro"]
              prioridad:
                type: "string"
                enum: ["Baja", "Media", "Alta", "Crítica"]
              ticket_id:
                type: "integer"
                example: 1
            required:
              - "descripcion"
              - "categoria"
              - "prioridad"
              - "ticket_id"
      responses:
        201:
          description: "Incidente creado exitosamente"
        400:
          description: "Parámetros inválidos"

  /incidentes/ticket/{ticket_id}:
    get:
      tags:
        - "Incidentes"
      summary: "Obtiene todos los incidentes de un ticket específico."
      parameters:
        - name: "ticket_id"
          in: "path"
          type: "integer"
          required: true
      responses:
        200:
          description: "Incidentes del ticket"

  /incidentes/{incidente_id}:
    get:
      tags:
        - "Incidentes"
      summary: "Obtiene los detalles de un incidente específico."
      parameters:
        - name: "incidente_id"
          in: "path"
          type: "integer"
          required: true
      responses:
        200:
          description: "Incidente encontrado"
        404:
          description: "Incidente no encontrado"
    delete:
      tags:
        - "Incidentes"
      summary: "Elimina un incidente."
      parameters:
        - name: "incidente_id"
          in: "path"
          type: "integer"
          required: true
      responses:
        200:
          description: "Incidente eliminado exitosamente"
        404:
          description: "Incidente no encontrado"

  /incidentes/filtrar/categoria:
    get:
      tags:
        - "Incidentes"
      summary: "Filtra incidentes por categoría."
      parameters:
        - name: "categoria"
          in: "query"
          type: "string"
          required: true
          enum: ["Hardware", "Software", "Red", "Otro"]
      responses:
        200:
          description: "Incidentes filtrados exitosamente"

  /incidentes/filtrar/prioridad:
    get:
      tags:
        - "Incidentes"
      summary: "Filtra incidentes por prioridad."
      parameters:
        - name: "prioridad"
          in: "query"
          type: "string"
          required: true
          enum: ["Baja", "Media", "Alta", "Crítica"]
      responses:
        200:
          description: "Incidentes filtrados exitosamente"